添加了操作数栈和本地变量表的打印函数，在MSCV中，long long 整形才是64位的，long整形是32位的。

+++ Gava/typeTrabslate.h
@@ -11,7 +11,7 @@ typedef union {
 
 typedef union {
 	u4 U4[2];
-	long l;
+	long long l;
 	double d;
 }stackType2;
